server_side_rendering_react/SSR-ts/STUB_SERVER_DYSON_CLASSES/src/helpers/getModalAddresses.ts
import { SavedAddress } from 'types/reducers'

export const getModalAddresses = (unsortedAddresses: SavedAddress[], modalRoute?: string) => {
  const nonContactAddresses = unsortedAddresses.filter((address) => !address.contactAddress)

  const contactAddress = unsortedAddresses.find((address) => address.contactAddress)
  if (typeof contactAddress === 'undefined') {
    return nonContactAddresses
  }

  return [contactAddress, ...nonContactAddresses].filter(
    (address) => address.lines.country === 'United Kingdom' || modalRoute !== 'delivery',
  )
}
